+++ R/ggplot_vbar.R
@@ -118,7 +118,7 @@ theme_vbar <-
 #' @param pal Character vector of hex codes. Defaults to NULL, which selects a default palette.
 #' @param width Width of bars. Defaults to 0.75.
 #' @param alpha The alpha of the fill. Defaults to 1. 
-#' @param line_size The size of the outlines of bars.
+#' @param size_line The size of the outlines of bars.
 #' @param title Title string. Defaults to [Title].
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. Not applicable where isMobile equals TRUE.
 #' @param subtitle Subtitle string. Defaults to [Subtitle].
@@ -165,7 +165,7 @@ ggplot_vbar <- function(data,
                         pal = NULL,
                         width = 0.75, 
                         alpha = 1,
-                        line_size = 0.5,
+                        size_line = 0.5,
                         title = "[Title]",
                         title_wrap = 70,
                         subtitle = NULL,
@@ -237,7 +237,7 @@ ggplot_vbar <- function(data,
       font_size_body = font_size_body,
       font_size_title = font_size_title
     ) +
-    geom_col(aes(x = !!x_var, y = !!y_var, text = !!text_var), col = pal, fill = pal, alpha = alpha, size = line_size, width = bar_width)
+    geom_col(aes(x = !!x_var, y = !!y_var, text = !!text_var), col = pal, fill = pal, alpha = alpha, size = size_line, width = bar_width)
   
   if (lubridate::is.Date(x_var_vctr) | is.numeric(x_var_vctr)) {
     
@@ -318,13 +318,13 @@ ggplot_vbar <- function(data,
       if(y_limits[1] >= 0 & y_limits[2] > 0){
         plot <- plot +
           geom_col(aes(x = !!y_var, y = y_limits[2], text = !!text_var),
-                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                    data = na_data)
       }
       else if(y_limits[1] < 0 & y_limits[2] <= 0) {
         plot <- plot +
           geom_col(aes(x = !!y_var, y = y_limits[1], text = !!text_var),
-                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                    data = na_data)        
       }
       else if(y_limits[1] < 0 & y_limits[2] > 0) {
@@ -332,10 +332,10 @@ ggplot_vbar <- function(data,
         
         plot <- plot +
           geom_col(aes(x = !!y_var, y = y_limits[2], text = !!text_var),
-                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                    data = na_data) +
           geom_col(aes(x = !!y_var, y = y_limits[1] + ggplotly_adjust, text = !!text_var),
-                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                   col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                    data = na_data)
       }
     }
@@ -385,7 +385,7 @@ ggplot_vbar <- function(data,
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param width Width of bars. Defaults to 0.75.
 #' @param alpha The alpha of the fill. Defaults to 1. 
-#' @param line_size The size of the outlines of bars.
+#' @param size_line The size of the outlines of bars.
 #' @param title Title string. Defaults to [Title].
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. Not applicable where isMobile equals TRUE.
 #' @param subtitle Subtitle string. Defaults to [Subtitle].
@@ -439,7 +439,7 @@ ggplot_vbar_col <-
            pal_rev = FALSE,
            width = 0.75, 
            alpha = 1,
-           line_size = 0.5,
+           size_line = 0.5,
            title = "[Title]",
            title_wrap = 70,
            subtitle = NULL,
@@ -536,7 +536,7 @@ ggplot_vbar_col <-
         font_size_title = font_size_title
       ) +
       geom_col(aes(x = !!x_var, y = !!y_var, col = !!col_var, fill = !!col_var, text = !!text_var), 
-               alpha = alpha, size = line_size, width = bar_width, position = position2)
+               alpha = alpha, size = size_line, width = bar_width, position = position2)
     
     if (lubridate::is.Date(x_var_vctr) | is.numeric(x_var_vctr)) {
       
@@ -622,7 +622,7 @@ ggplot_vbar_col <-
       plot <- plot +
         geom_col(aes(
           x = !!x_var, y = !!y_var, col = !!col_var, fill = !!col_var, text = !!text_var), 
-          alpha = alpha, size = line_size, width = width, 
+          alpha = alpha, size = size_line, width = width, 
           position = position2)
     }
     else if(y_na_inf == TRUE) {
@@ -649,7 +649,7 @@ ggplot_vbar_col <-
         
         plot <- plot +
           geom_col(aes(x = !!x_var, y = .data$y_var2, col = .data$col_var2, fill = .data$col_var2, group = !!col_var, text = !!text_var), 
-                   alpha = alpha, size = line_size, width = width, position = position2, data = data)
+                   alpha = alpha, size = size_line, width = width, position = position2, data = data)
       }
       else if(y_limits[1] < 0 & y_limits[2] <= 0) {
         data <- data %>%
@@ -657,7 +657,7 @@ ggplot_vbar_col <-
         
         plot <- plot +
           geom_col(aes(x = !!x_var, y = .data$y_var2, col = .data$col_var2, fill = .data$col_var2, group = !!col_var, text = !!text_var), 
-                   alpha = alpha, size = line_size, width = width, position = position2, data = data)
+                   alpha = alpha, size = size_line, width = width, position = position2, data = data)
       }
       else if(y_limits[1] < 0 & y_limits[2] > 0) {
         data <- data %>%
@@ -667,9 +667,9 @@ ggplot_vbar_col <-
         
         plot <- plot +
           geom_col(aes(x = !!x_var, y = .data$y_var2, col = .data$col_var2, fill = .data$col_var2, group = !!col_var, text = !!text_var), 
-                   alpha = alpha, size = line_size, width = width, position = position2, data = data) +
+                   alpha = alpha, size = size_line, width = width, position = position2, data = data) +
           geom_col(aes(x = !!x_var, y = .data$y_var3, col = .data$col_var2, fill = .data$col_var2, group = !!col_var, text = !!text_var), 
-                   alpha = alpha, size = line_size, width = width, position = position2, data = data)
+                   alpha = alpha, size = size_line, width = width, position = position2, data = data)
       }
     }
     
@@ -751,7 +751,7 @@ ggplot_vbar_col <-
 #' @param pal Character vector of hex codes. Defaults to NULL, which selects a default palette.
 #' @param width Width of bars. Defaults to 0.75.
 #' @param alpha The alpha of the fill. Defaults to 1.
-#' @param line_size The size of the outlines of bars. 
+#' @param size_line The size of the outlines of bars. 
 #' @param title Title string. Defaults to [Title].
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. 
 #' @param subtitle Subtitle string. Defaults to [Subtitle].
@@ -800,7 +800,7 @@ ggplot_vbar_facet <-
            pal = NULL,
            width = 0.75, 
            alpha = 1,
-           line_size = 0.5,
+           size_line = 0.5,
            title = "[Title]",
            title_wrap = 70,
            subtitle = NULL,
@@ -871,7 +871,7 @@ ggplot_vbar_facet <-
         font_size_body = font_size_body,
         font_size_title = font_size_title
       ) +
-      geom_col(aes(x = !!x_var, y = !!y_var, text = !!text_var), col = pal, fill = pal, alpha = alpha, size = line_size, width = bar_width)
+      geom_col(aes(x = !!x_var, y = !!y_var, text = !!text_var), col = pal, fill = pal, alpha = alpha, size = size_line, width = bar_width)
     
     if (facet_scales %in% c("fixed", "free_y")) {
       if (lubridate::is.Date(x_var_vctr) | is.numeric(x_var_vctr)) {
@@ -945,13 +945,13 @@ ggplot_vbar_facet <-
           if(y_limits[1] >= 0 & y_limits[2] > 0){
             plot <- plot +
               geom_col(aes(x = !!y_var, y = y_limits[2], text = !!text_var),
-                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                        data = na_data)
           }
           else if(y_limits[1] < 0 & y_limits[2] <= 0) {
             plot <- plot +
               geom_col(aes(x = !!y_var, y = y_limits[1], text = !!text_var),
-                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                        data = na_data)        
           }
           else if(y_limits[1] < 0 & y_limits[2] > 0) {
@@ -959,10 +959,10 @@ ggplot_vbar_facet <-
             
             plot <- plot +
               geom_col(aes(x = !!y_var, y = y_limits[2], text = !!text_var),
-                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                       col = "#F5F5F5", fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                        data = na_data) +
               geom_col(aes(x = !!y_var, y = y_limits[1] + ggplotly_adjust, text = !!text_var),
-                       fill = "#F5F5F5", alpha = alpha, size = line_size, width = width, 
+                       fill = "#F5F5F5", alpha = alpha, size = size_line, width = width, 
                        data = na_data)
           }
         }
@@ -1008,7 +1008,7 @@ ggplot_vbar_facet <-
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param width Width of bars. Defaults to 0.75.
 #' @param alpha The alpha of the fill. Defaults to 1. 
-#' @param line_size The size of the outlines of bars.
+#' @param size_line The size of the outlines of bars.
 #' @param title Title string. Defaults to [Title].
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 70. 
 #' @param subtitle Subtitle string. Defaults to [Subtitle].
@@ -1068,7 +1068,7 @@ ggplot_vbar_col_facet <-
            pal_rev = FALSE,
            width = 0.75, 
            alpha = 1,
-           line_size = 0.5,
+           size_line = 0.5,
            title = "[Title]",
            title_wrap = 70,
            subtitle = NULL,
@@ -1163,7 +1163,7 @@ ggplot_vbar_col_facet <-
         font_size_title = font_size_title
       ) +
       geom_col(aes(x = !!x_var, y = !!y_var, col = !!col_var, fill = !!col_var, text = !!text_var), 
-               alpha = alpha, size = line_size, width = bar_width, position = position2)
+               alpha = alpha, size = size_line, width = bar_width, position = position2)
     
     if (position == "stack") {
       data_sum <- data %>%